Moving log related functions to separate module
